fix(query-api): scan ClickHouse aggregates at real UInt64 type (#2619)
## Summary

sankey and investmentflow read several ClickHouse aggregates
(`sum()`, `countIf()`, `countDistinctIf()`) over UInt32-width source
columns and scan the results into Go destinations narrower than what
the aggregate actually returns. ClickHouse promotes every one of these
aggregates to UInt64 regardless of the source column's own width, and
the driver refuses to scan a UInt64 result into `*float64` or `*int64`
-- this degrades the affected reads to a 503 in production.

The fix follows what the reference plane's service layer does to each
value before it reaches the wire, not a uniform rule:

- Fields the reference wraps in `float(...)` before use (`new_items`,
  `new_bugs`, `items_touched`, `churn`, `canceled_items`) now `CAST` to
  `Float64` in the SQL itself, keeping a single float64 path end to
  end and matching that float wire contract.
- Fields the reference wraps in `int(...)` before use (`missing_repo`,
  `missing_team`) now scan into `uint64` directly, matching the
  driver's actual result type and the plain-integer wire contract
  those two fields already have.

`canceled_items` (the expense-abandoned read) was hidden behind an
earlier read's own failure in the same flow -- fixing that earlier
read's scan type is what let this one execute for the first time and
surface the same class of bug underneath it.

Every fixture-backed test in both packages answers a `Scan` call by
reflecting on the Go destination type it is handed, not by replaying
the real driver's own conversion rules, so none of them exercised this
path. Both packages gain a real-ClickHouse (testcontainers) recurrence
guard that scans these exact aggregate shapes against a seeded schema.
The CI integration-package inventory and shard manifest are updated
for both packages' first `-tags=integration` file.

Every 503 degradation site already logs its wrapped cause with route,
org id and request id; that telemetry is unchanged by this change and
is what made this diagnosable.

## TEST-EVIDENCE

`go build ./...`
```
(clean, exit 0)
```

`go vet ./cmd/query-api/... ./internal/migrationmatrix/...`
```
(clean, exit 0)
```

`go test ./cmd/query-api/... ./internal/migrationmatrix/...`
```
ok  	github.com/full-chaos/dev-health-ops/cmd/query-api	3.197s
ok  	github.com/full-chaos/dev-health-ops/cmd/query-api/internal/investmentflow	0.022s
ok  	github.com/full-chaos/dev-health-ops/cmd/query-api/internal/sankey	0.016s
ok  	github.com/full-chaos/dev-health-ops/internal/migrationmatrix	0.095s
(plus every other cmd/query-api/internal/* package, all ok)
```

`go test -tags=integration ./cmd/query-api/internal/sankey/... ./cmd/query-api/internal/investmentflow/...`
(real ClickHouse via testcontainers)
```
ok  	github.com/full-chaos/dev-health-ops/cmd/query-api/internal/sankey	12.186s
ok  	github.com/full-chaos/dev-health-ops/cmd/query-api/internal/investmentflow	11.584s
```
Confirmed each new integration test fails against the pre-fix scan
types with the same driver error class captured in production, and
passes against the fix.

The seeded DDL for `work_item_metrics_daily`, `work_item_state_durations_daily`
and `file_metrics_daily` declares `ReplacingMergeTree(computed_at)` with
each table's real `ORDER BY`, taken from a confirmed live read of prod
`system.tables` (not from this repo's migration files, which still show
a plain `MergeTree` for all three with no later engine-changing
migration -- a live schema/migration mismatch, unrelated to this fix and
not touched here).

`uv run pytest tests/tooling/test_go_integration_sharding.py tests/tooling/test_check_go_integration_coverage.py tests/tooling/test_go_workflow_path_filters.py`
```
93 passed
```
Package counts (`63 package(s) discovered`, shard-plan `63 package(s)`,
`62` for the providersync-excluded set) were read directly off
`ci/check_go.sh integration-coverage` / `integration-shard-plan` run
against the current tree, not derived by arithmetic.

Golden fixtures: `testdata/*.json` in both packages are untouched --
every `TestGolden*` test still passes against the same committed
golden files, so output is unchanged. Only the in-memory fixture rows
that feed those tests (previously typed to bypass the real driver's
own conversion rules) changed type to match what the fix now scans.

## RISK-NOTES

- No declared behavioral divergence from the reference plane: every
  changed field keeps the same numeric value on the wire, only the
  ClickHouse-side type and the Go scan destination change to match
  what the driver actually returns.
- Schema note for future readers: this package reads
  `work_item_metrics_daily`, `work_item_state_durations_daily` and
  `file_metrics_daily` with `FINAL`, which only a dedup-capable engine
  (e.g. ReplacingMergeTree) supports. A live read of prod `system.tables`
  confirms all three are `ReplacingMergeTree(computed_at)`, the same
  shape two sibling tables' own later migrations already moved to --
  while the checked-in migration history for these three tables shows
  only the original plain `MergeTree` create, with no later
  engine-changing migration. The live schema for these three tables has
  drifted from what the migration files show. This PR does not touch
  those `FINAL` clauses (they predate this change) and does not attempt
  the reconciliation; flagging so a future reader does not trust the
  migration file over live behavior here.
- No rate-limiter gap and no default-off toggle are introduced by this
  change.

Ticket: CHAOS-5866

+++ cmd/query-api/internal/sankey/golden_test.go
@@ -102,7 +102,7 @@ func TestGoldenExpense(t *testing.T) {
 		case strings.Contains(query, "FROM work_item_metrics_daily FINAL"):
 			return &fixtureRowScanner{rows: [][]any{{10.0, 8.0, 5.0}}}, nil
 		case strings.Contains(query, "FROM work_item_cycle_times FINAL"):
-			return &fixtureRowScanner{rows: [][]any{{int64(2)}}}, nil
+			return &fixtureRowScanner{rows: [][]any{{2.0}}}, nil
 		default:
 			t.Fatalf("unexpected query for expense golden:\n%s", query)
 			return nil, nil
